LPS-141069 Move to proper package
