+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-forms-web/src/main/resources/META-INF/resources/admin/init.jsp
@@ -53,7 +53,7 @@ page import="com.liferay.portal.kernel.workflow.WorkflowDefinition" %><%@
 page import="com.liferay.portal.kernel.workflow.WorkflowDefinitionManagerUtil" %><%@
 page import="com.liferay.portal.kernel.workflow.WorkflowTask" %><%@
 page import="com.liferay.portal.kernel.workflow.WorkflowTaskManagerUtil" %><%@
-page import="com.liferay.portal.workflow.instance.tracker.util.InstanceTrackerURLProviderUtil" %><%@
+page import="com.liferay.portal.workflow.instance.tracker.url.provider.InstanceTrackerURLProviderUtil" %><%@
 page import="com.liferay.portal.workflow.kaleo.designer.web.constants.KaleoDesignerPortletKeys" %><%@
 page import="com.liferay.portal.workflow.kaleo.forms.exception.KaleoProcessDDMTemplateIdException" %><%@
 page import="com.liferay.portal.workflow.kaleo.forms.model.KaleoProcessLink" %><%@
